Rename year and quarter date range data classes
