+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/ranges/data/LastYearRangeData.kt
@@ -1,5 +1,6 @@
 package com.woocommerce.android.ui.analytics.ranges.data
 
+import com.woocommerce.android.extensions.endOfCurrentYear
 import com.woocommerce.android.extensions.oneYearAgo
 import com.woocommerce.android.extensions.startOfCurrentYear
 import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubTimeRange
@@ -7,7 +8,7 @@ import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubTimeRangeData
 import java.util.Calendar
 import java.util.Date
 
-class AnalyticsHubYearToDateRangeData(
+class LastYearRangeData(
     referenceDate: Date,
     referenceCalendar: Calendar
 ) : AnalyticsHubTimeRangeData(referenceCalendar) {
@@ -15,17 +16,18 @@ class AnalyticsHubYearToDateRangeData(
     override val previousRange: AnalyticsHubTimeRange
 
     init {
-        calendar.time = referenceDate
-        currentRange = AnalyticsHubTimeRange(
-            start = calendar.startOfCurrentYear(),
-            end = referenceDate
-        )
-
         val oneYearAgo = referenceDate.oneYearAgo(calendar)
         calendar.time = oneYearAgo
+        currentRange = AnalyticsHubTimeRange(
+            start = calendar.startOfCurrentYear(),
+            end = calendar.endOfCurrentYear()
+        )
+
+        val twoYearsAgo = oneYearAgo.oneYearAgo(calendar)
+        calendar.time = twoYearsAgo
         previousRange = AnalyticsHubTimeRange(
             start = calendar.startOfCurrentYear(),
-            end = oneYearAgo
+            end = calendar.endOfCurrentYear()
         )
     }
 }
